Refactors the file structure
